+++ front-server/src/store/modules/movies/comment.js
@@ -1,7 +1,7 @@
 import axios from 'axios'
 // import router from 'vue-router'
 // 
-import { API_URL } from '@/store/CONSTS'
+import { API_URL } from '@/store/modules/movies/CONSTS'
 
 const state = () => {
   return {
